requirements.txt
astroid==1.5.3
bleach==1.5.0
colorama==0.3.9
cycler==0.10.0
dask==0.15.4
decorator==4.1.2
Django==1.11.7
django-braces==1.11.0
django-cleanup==1.0.1
django-extensions==1.9.7
djangorestframework==3.7.3
entrypoints==0.2.3
enum34==1.1.6
h5py==2.7.1
html5lib==0.9999999
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.3
isort==4.2.15
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.0.9
lazy-object-proxy==1.3.1
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
matplotlib-colorbar==0.3.5
mccabe==0.6.1
mistune==0.8.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
notebook==5.2.1
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
pandocfilters==1.4.2
parso==0.1.1
pickleshare==0.7.4
Pillow==4.3.0
ply==3.9.0
prompt-toolkit==1.0.15
protobuf==3.4.0
psycopg2==2.7.3.2
Pygments==2.2.0
pylint==1.7.4
pyparsing==2.2.0
python-dateutil==2.6.1
python-pcapng==0.2
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
scikit-image==0.13.0
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
testpath==0.3.1
toolz==0.8.2
tornado==4.5.2
traitlets==4.3.2
typing==3.6.2
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.6
win-unicode-console==0.5
wrapt==1.10.11
django-multiupload==0.5.2
django-webpack-loader==0.5.0
django-widget-tweaks==1.4.1
django-filter==1.1.0